backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary  # needed only for PostgreSQL
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
Flask
Flask-SQLAlchemy
flask-jwt-extended
psycopg2-binary
Werkzeug
python-dotenv
PyJWT
